backend/requirements.txt
flask==3.0.2
flask-socketio==5.3.6
eventlet==0.35.1
flask-cors==4.0.0
pyyaml==6.0.1